Add `onStartShouldSetPanResponder: () => true` so animation starts on press down.
Without this, the user needs to start moving the object for the scale transform to occur. 

In my experience, pressing down on an object and having that object "pop out" of the page (via the scale transform) is a really nice indicator that we are now dragging the object. 

Adding `onStartShouldSetPanResponder: () => true` does just that.